pyproject.toml
[project]
name = "basic_chat_configurable"
version = "0.1.0"
description = "Single system prompt voice agent that uses Gemini 2.5 Flash"
requires-python = ">=3.10"

dependencies = [
    "cartesia-line",
    "google-genai>=1.26.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.12.0",
    "uvicorn==0.35.0",
    "langchain-core",
    "langchain-google-genai",
    "supabase",
    "asyncio",
    "twilio",
    ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]


[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist==3.8.0",
    "pytest-repeat==0.9.4",
    "google-genai>=1.26.0,<2",
]
